Update project - Recycler View - refactoring
