+++ app/src/main/java/com/example/moves/data/database/MovieDatabase.kt
@@ -4,6 +4,7 @@ import android.app.Application
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.moves.data.database.model.MovieDbModel
 
 @Database(entities = [MovieDbModel::class], version = 1, exportSchema = false)
 abstract class MovieDatabase : RoomDatabase() {
