--- app/src/main/java/com/example/moves/data/database/PosterDb.kt
@@ -1,3 +1,0 @@
-package com.example.moves.data.database
-
-class PosterDb(val url: String)
