--- app/src/main/java/com/example/moves/data/database/RatingDb.kt
@@ -1,5 +1,0 @@
-package com.example.moves.data.database
-
-import com.google.gson.annotations.SerializedName
-
-class RatingDb(@SerializedName("kp") val kp: Double)
