+++ app/src/main/java/com/example/moves/data/database/model/PosterDbModel.kt
@@ -1,0 +1,3 @@
+package com.example.moves.data.database.model
+
+class PosterDbModel(val url: String)
