+++ app/src/main/java/com/example/moves/data/database/model/RatingDbModel.kt
@@ -1,0 +1,5 @@
+package com.example.moves.data.database.model
+
+import com.google.gson.annotations.SerializedName
+
+class RatingDbModel(@SerializedName("kp") val kp: Double)
